requirements.txt
ollama
requests
fastapi
pydantic
uvicorn
pyyaml
langchain
langchain-core
langchain-ollama
langchain-community
langgraph
pytest
